Versie 0.7.2 - Translations gefixed, kleine wijzigingen n.a.v. demo 6-5: bankrekening gefixed, rekenregel voor 2025 gefixed (KPI6), files downloaden ipv openen, hotfix DZHM alleen voor 2021, ANlb neemt ook langere tags mee dan 4 chars, Xls/Xlsx ondersteuning (werkt niet goed), dumps ingekort, support voor langere KVK-nummers, check op geldige IBAN in Laravel en React, volgorde in Admin-menu aangepast)

+++ database/migrations/create_gis_records_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('gis_records', function (Blueprint $table) {
             $table->id();
             $table->foreignId('dump_id');
-            $table->integer('kvk');
+            $table->bigInteger('kvk');
             $table->string('eenheid_code');
             $table->float('lengte');
             $table->float('breedte');
